Move `DatabaseCustomError` to `astro/exceptions.py`
Looks like we already had a custom exceptions file with 2 custom exceptions. So just making it consistent and moving `DatabaseCustomError` to this file
