+++ tests/databases/test_snowflake.py
@@ -11,9 +11,8 @@ from sqlalchemy.exc import ProgrammingError
 
 from astro.constants import Database, FileLocation, FileType
 from astro.databases import create_database
-from astro.databases.base import DatabaseCustomError
 from astro.databases.snowflake import SnowflakeDatabase, SnowflakeStage
-from astro.exceptions import NonExistentTableException
+from astro.exceptions import DatabaseCustomError, NonExistentTableException
 from astro.files import File
 from astro.settings import (
     SCHEMA,
